requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20230802.0=h6a678d5_2
absl-py=1.4.0=py38h06a4308_0
aiohttp=3.8.5=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
alembic=1.8.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.2=py38h06a4308_0
attrs=23.1.0=py38h06a4308_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
blinker=1.6.2=py38h06a4308_0
bottle=0.12.25=pypi_0
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py38h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py38h06a4308_0
cloudpickle=2.2.1=py38h06a4308_0
colorlog=6.7.0=py38h578d9bd_2
contourpy=1.0.5=py38hdb19cb5_0
cryptography=41.0.3=py38hdda0065_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.52=0
cuda-runtime=12.1.0=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
dpkt=1.9.8=pyhd8ed1ab_0
einops=0.7.0=pyhd8ed1ab_1
entrypoints=0.4=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py38h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py38h5eee18b_0
fsspec=2023.10.0=py38h06a4308_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=py38h06a4308_0
google-auth-oauthlib=0.5.2=py38h06a4308_0
greenlet=2.0.1=py38h6a678d5_0
grpc-cpp=1.48.2=he1ff14a_3
grpcio=1.48.2=py38he1ff14a_3
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gtest=1.14.0=hdb19cb5_0
huggingface_hub=0.20.3=py38h06a4308_0
icu=73.1=h6a678d5_0
idna=3.4=py38h06a4308_0
importlib-metadata=6.0.0=py38h06a4308_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=6.1.1=pyhd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.14.0=py38h7f3c49e_0
ipython=8.4.0=py38h578d9bd_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.2.0=py38h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=5.5.0=py38h578d9bd_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.8.0.34=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.4.52=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
libxslt=1.1.37=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.0=py38he621ea3_0
lxml=4.9.3=py38hdbbb534_0
lz4-c=1.9.4=h6a678d5_0
mako=1.3.0=pyhd8ed1ab_0
markdown=3.4.1=py38h06a4308_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multidict=6.0.2=py38h5eee18b_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
numba=0.58.0=py38h1128e8f_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=py38h06a4308_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
optuna=3.4.0=pyhd8ed1ab_0
optuna-dashboard=0.13.0=pypi_0
packaging=23.1=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
ply=3.11=py38_0
pooch=1.7.0=py38h06a4308_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=3.20.3=py38h6a678d5_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pyshark=0.6=pyhd8ed1ab_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.1.0=py3.8_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-metric-learning=2.3.0=pyhd8ed1ab_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py38h06a4308_0
pyyaml=6.0.1=py38h5eee18b_0
pyzmq=19.0.2=py38ha71036d_2
qt-main=5.15.2=h53bd1ea_10
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.4.2=py38ha89cbab_0
scienceplots=2.1.0=pypi_0
scikit-learn=1.3.0=py38h1128e8f_0
scipy=1.10.1=py38hf6e8229_1
seaborn=0.12.2=py38h06a4308_0
setuptools=68.0.0=py38h06a4308_0
shap=0.42.1=py38h1128e8f_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
slicer=0.0.7=pyhd3eb1b0_0
sqlalchemy=1.4.39=py38h5eee18b_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.1=py38h06a4308_0
tensorboard-data-server=0.7.0=py38h52d8a92_0
tensorboard-plugin-wit=1.8.1=py38h06a4308_0
termcolor=2.3.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=0.9.16=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py38h06a4308_0
torchaudio=2.1.0=py38_cu121
torchtriton=2.1.0=py38
torchvision=0.16.0=py38_cu121
tornado=6.3.3=py38h5eee18b_0
tqdm=4.65.0=py38hb070fc8_0
traitlets=5.14.0=pyhd8ed1ab_0
typing-extensions=4.9.0=py38h06a4308_1
typing_extensions=4.9.0=py38h06a4308_1
urllib3=1.26.18=py38h06a4308_0
wcwidth=0.2.12=pyhd8ed1ab_0
werkzeug=2.2.3=py38h06a4308_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py38h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
